+++ include/mola_state_estimation_smoother/StateEstimationSmoother.h
@@ -27,9 +27,9 @@
 #include <mola_kernel/interfaces/NavStateFilter.h>
 #include <mola_kernel/interfaces/RawDataSourceBase.h>
 #include <mola_kernel/interfaces/VizInterface.h>
+#include <mola_kernel/utils/RegexCache.h>
 #include <mola_kernel/version.h>
 #include <mola_state_estimation_smoother/Parameters.h>
-#include <mola_state_estimation_smoother/RegexCache.h>
 
 // MOLA:
 #include <mola_imu_preintegration/ImuIntegrator.h>
@@ -234,6 +234,9 @@ class StateEstimationSmoother : public mola::NavStateFilter,
 
         /// A bimap of known odometry "frame_id" <=> "numeric IDs":
         mrpt::containers::bimap<std::string, odometry_frameid_t> known_odom_frames;
+
+        /// Monotonic counter for odometry frame IDs; avoids collisions if entries are ever removed.
+        odometry_frameid_t next_odom_frame_id = 1;
 
         /// The latest values from the estimator; updated in process_pending_gtsam_updates()
         std::map<frame_index_t, FrameState> last_estimated_states;